+++ packages/adapter-kaiheila/src/bot.ts
@@ -45,10 +45,10 @@ export class KaiheilaBot extends Bot {
     return result.data
   }
 
-  private _prepareHandle(channelId: string, content: string): SendHandle {
+  private _prepareHandle(channelId: string, content: string, groupId: string): SendHandle {
     let path: string
     const params = {} as KHL.MessageParams
-    const session = this.createSession({ channelId, content })
+    const session = this.createSession({ channelId, content, groupId })
     if (channelId.length > 30) {
       params.chatCode = channelId
       session.subtype = 'private'
@@ -56,8 +56,6 @@ export class KaiheilaBot extends Bot {
     } else {
       params.targetId = channelId
       session.subtype = 'group'
-      // FIXME this is incorrect but to workarournd ctx.group()
-      session.groupId = 'unknown'
       path = '/message/create'
     }
     return [path, params, session]
@@ -92,15 +90,16 @@ export class KaiheilaBot extends Bot {
     }
   }
 
-  private async _sendCard(handle: SendHandle, chain: segment.Chain) {
-    let text: KHL.Card.Text = { type: 'plain-text', content: '' }
+  private async _sendCard(handle: SendHandle, chain: segment.Chain, useMarkdown: boolean) {
+    const type = useMarkdown ? 'kmarkdown' : 'plain-text'
+    let text: KHL.Card.Text = { type, content: '' }
     let card: KHL.Card = { type: 'card', modules: [] }
     const output: KHL.Card[] = []
     const flushText = () => {
       text.content = text.content.trim()
       if (!text.content) return
       card.modules.push({ type: 'section', text })
-      text = { type: 'plain-text', content: '' }
+      text = { type, content: '' }
     }
     const flushCard = () => {
       flushText()
@@ -150,12 +149,13 @@ export class KaiheilaBot extends Bot {
     await this._sendHandle(handle, KHL.Type.card, JSON.stringify(output))
   }
 
-  private async _sendSeparate(handle: SendHandle, chain: segment.Chain) {
+  private async _sendSeparate(handle: SendHandle, chain: segment.Chain, useMarkdown: boolean) {
     let textBuffer = ''
+    const type = useMarkdown ? KHL.Type.kmarkdown : KHL.Type.text
     const flush = async () => {
       textBuffer = textBuffer.trim()
       if (!textBuffer) return
-      await this._sendHandle(handle, KHL.Type.text, textBuffer)
+      await this._sendHandle(handle, type, textBuffer)
       handle[1].quote = null
       textBuffer = ''
     }
@@ -187,14 +187,19 @@ export class KaiheilaBot extends Bot {
     await flush()
   }
 
-  async sendMessage(channelId: string, content: string) {
-    const handle = this._prepareHandle(channelId, content)
+  async sendMessage(channelId: string, content: string, groupId?: string) {
+    const handle = this._prepareHandle(channelId, content, groupId)
     const [, params, session] = handle
     if (await this.app.serial(session, 'before-send', session)) return
 
+    let useMarkdown = false
     const chain = segment.parse(content)
     if (chain[0].type === 'quote') {
       params.quote = chain.shift().data.id
+    }
+    if (chain[0].type === 'markdown') {
+      useMarkdown = true
+      chain.shift()
     }
 
     const { attachMode } = this.app.options.kaiheila
@@ -202,9 +207,9 @@ export class KaiheilaBot extends Bot {
     const useCard = hasAttachment && (attachMode === 'card' || attachMode === 'mixed' && chain.length > 1)
 
     if (useCard) {
-      await this._sendCard(handle, chain)
+      await this._sendCard(handle, chain, useMarkdown)
     } else {
-      await this._sendSeparate(handle, chain)
+      await this._sendSeparate(handle, chain, useMarkdown)
     }
 
     return session.messageId